pyproject.toml
[tool.poetry]
name = "scrap-products"
version = "0.1.0"
description = ""
authors = ["Vojd <regush20@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
selenium = "^4.38.0"
requests = "^2.32.5"
bs4 = "^0.0.2"
webdriver-manager = "^4.0.2"
python-dotenv = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
